test: 대회 취소 기반 비동기 알림 통합 테스트 (#100)
* test: 비동기 알람 리스너 동작 검증용 통합 테스트 추가 (#93)

* test: 트랜잭션 롤백 시 이메일 미발송 통합 테스트 추가

* test: 이메일 장애 시 접수 데이터 정상 커밋 검증

* test: 참가 접수 성공 시 이메일 비동기 발송 검증

* refactor: 통합 테스트 데이터 중복 제거 (#93)

* test: 대회 취소 이벤트 커밋/롤백 통합 테스트 추가 (#100)

* test: 대회 취소 성공 시 이메일 발송 통합 테스트 추가 (#100)

* test: 대회 취소 이메일 실패 시 트랜잭션 유지 검증 추가 (#100)
